+++ config/dep_all.rs
@@ -1,5 +1,5 @@
-// In sync with ../Cargo.toml::dep_all, ../build/features.rs::DEPENDENCY & ../src/_dep.rs
-// Used from: ../build/features.rs::DEPENDENCY.features & ../tools/check.rs::DEP_ALL.
+// In sync with ../Cargo.toml::dep_all, ../build/main/features.rs::DEPENDENCY & ../src/_dep.rs
+// Used from: ../build/main/features.rs::DEPENDENCY.features & ../tools/check.rs::DEP_ALL.
 //
 // NOTE: make sure to recompile `../tools/check.rs` after modifying this file.
 // 44 optional dependencies: