+++ config/lib_all.rs
@@ -1,5 +1,5 @@
 // In sync with ../Cargo.toml::lib_all
-// Used from: ../build/features.rs::LIBRARY.features & ../tools/check.rs::LIB_ALL.
+// Used from: ../build/main/features.rs::LIBRARY.features & ../tools/check.rs::LIB_ALL.
 //
 // NOTE: make sure to recompile `../tools/check.rs` after modifying this file.
 // 0 optional dependencie: